refactor(core/search): Rename *SearchResultType files
