+++ app/scripts/modules/core/src/application/applicationSearchResultType.ts
