+++ app/scripts/modules/core/src/cluster/clusterSearchResultType.ts
