+++ app/scripts/modules/core/src/instance/instanceSearchResultType.ts
