+++ app/scripts/modules/core/src/loadBalancer/loadBalancer.module.ts
@@ -3,7 +3,7 @@ import { module } from 'angular';
 import { LOAD_BALANCER_DATA_SOURCE } from './loadBalancer.dataSource';
 import { LOAD_BALANCER_STATES } from './loadBalancer.states';
 import { LOAD_BALANCER_FILTER } from './filter/loadBalancer.filter.component';
-import './LoadBalancerSearchResultFormatter';
+import './loadBalancerSearchResultType';
 
 export const LOAD_BALANCER_MODULE = 'spinnaker.core.loadBalancer';
