+++ app/scripts/modules/core/src/projects/projectSearchResultType.ts
