+++ app/scripts/modules/core/src/serverGroup/serverGroupSearchResultType.ts
